broke code out into smaller modules.
