+++ lib/helpers.js
@@ -1,0 +1,10 @@
+(function() {
+  var path;
+
+  path = require('path');
+
+  module.exports = {
+    getUrlPath: getUrlPath
+  };
+
+}).call(this);
