+++ lib/snockets.js
@@ -1,5 +1,5 @@
 (function() {
-  var CoffeeScript, DIRECTIVE, DepGraph, EXPLICIT_PATH, HEADER, HoldingQueue, Snockets, SourceMap, compilers, fs, getUrlPath, jsExts, minify, parseDirectives, path, sourceMapCat, stripExt, timeEq, uglify, _,
+  var DIRECTIVE, DepGraph, EXPLICIT_PATH, HEADER, HoldingQueue, Snockets, SourceMap, compilers, fs, getUrlPath, jsExts, minify, parseDirectives, path, sourceMapCat, stripExt, timeEq, _, _ref, _ref1,
     __bind = function(fn, me){ return function(){ return fn.apply(me, arguments); }; },
     __slice = [].slice,
     __indexOf = [].indexOf || function(item) { for (var i = 0, l = this.length; i < l; i++) { if (i in this && this[i] === item) return i; } return -1; };
@@ -8,41 +8,43 @@
 
   SourceMap = require('source-map');
 
-  CoffeeScript = require('coffee-script');
-
   fs = require('fs');
 
   path = require('path');
 
-  uglify = require('uglify-js');
-
   _ = require('underscore');
+
+  minify = require('./minification').minify;
+
+  _ref = require('./util'), timeEq = _ref.timeEq, getUrlPath = _ref.getUrlPath;
+
+  _ref1 = require('./compilers'), compilers = _ref1.compilers, jsExts = _ref1.jsExts, stripExt = _ref1.stripExt;
 
   module.exports = Snockets = (function() {
     function Snockets(options) {
-      var errorStr, _base, _base1, _base2, _base3, _base4, _base5, _ref, _ref1, _ref2, _ref3, _ref4, _ref5;
+      var errorStr, _base, _base1, _base2, _base3, _base4, _base5, _ref2, _ref3, _ref4, _ref5, _ref6, _ref7;
 
       this.options = options != null ? options : {};
       this.error = __bind(this.error, this);
       this.debug = __bind(this.debug, this);
       this.warn = __bind(this.warn, this);
       this.info = __bind(this.info, this);
-      if ((_ref = (_base = this.options).srcmap) == null) {
+      if ((_ref2 = (_base = this.options).srcmap) == null) {
         _base.srcmap = false;
       }
-      if ((_ref1 = (_base1 = this.options).target) == null) {
+      if ((_ref3 = (_base1 = this.options).target) == null) {
         _base1.target = null;
       }
-      if ((_ref2 = (_base2 = this.options).staticRoot) == null) {
+      if ((_ref4 = (_base2 = this.options).staticRoot) == null) {
         _base2.staticRoot = null;
       }
-      if ((_ref3 = (_base3 = this.options).staticRootUrl) == null) {
+      if ((_ref5 = (_base3 = this.options).staticRootUrl) == null) {
         _base3.staticRootUrl = '/';
       }
-      if ((_ref4 = (_base4 = this.options).src) == null) {
+      if ((_ref6 = (_base4 = this.options).src) == null) {
         _base4.src = '.';
       }
-      if ((_ref5 = (_base5 = this.options).async) == null) {
+      if ((_ref7 = (_base5 = this.options).async) == null) {
         _base5.async = true;
       }
       this.cache = {};
@@ -98,7 +100,7 @@
     };
 
     Snockets.prototype.scan = function(filePath, flags, callback) {
-      var _ref,
+      var _ref2,
         _this = this;
 
       if (typeof flags === 'function') {
@@ -108,7 +110,7 @@
       if (flags == null) {
         flags = {};
       }
-      if ((_ref = flags.async) == null) {
+      if ((_ref2 = flags.async) == null) {
         flags.async = this.options.async;
       }
       return this.updateDirectives(filePath, flags, function(err, graphChanged) {
@@ -127,7 +129,7 @@
     };
 
     Snockets.prototype.getCompiledChain = function(filePath, flags, callback) {
-      var _ref,
+      var _ref2,
         _this = this;
 
       if (typeof flags === 'function') {
@@ -137,7 +139,7 @@
       if (flags == null) {
         flags = {};
       }
-      if ((_ref = flags.async) == null) {
+      if ((_ref2 = flags.async) == null) {
         flags.async = this.options.async;
       }
       return this.updateDirectives(filePath, flags, function(err, graphChanged) {
@@ -161,12 +163,12 @@
           }
         }
         compiledChain = (function() {
-          var _i, _len, _ref1, _results;
+          var _i, _len, _ref3, _results;
 
-          _ref1 = chain.concat(filePath);
+          _ref3 = chain.concat(filePath);
           _results = [];
-          for (_i = 0, _len = _ref1.length; _i < _len; _i++) {
-            link = _ref1[_i];
+          for (_i = 0, _len = _ref3.length; _i < _len; _i++) {
+            link = _ref3[_i];
             o = {};
             if (this.compileFile(link)) {
               o.filename = stripExt(link) + '.js';
@@ -186,7 +188,7 @@
     };
 
     Snockets.prototype.getConcatenation = function(filePath, flags, callback) {
-      var concatenationChanged, _ref,
+      var concatenationChanged, _ref2,
         _this = this;
 
       if (typeof flags === 'function') {
@@ -196,7 +198,7 @@
       if (flags == null) {
         flags = {};
       }
-      if ((_ref = flags.async) == null) {
+      if ((_ref2 = flags.async) == null) {
         flags.async = this.options.async;
       }
       concatenationChanged = true;
@@ -417,9 +419,9 @@
       graphChanged = false;
       q = new HoldingQueue({
         task: function(depPath, next) {
-          var err, _ref;
+          var err, _ref2;
 
-          if (_ref = path.extname(depPath), __indexOf.call(jsExts(), _ref) < 0) {
+          if (_ref2 = path.extname(depPath), __indexOf.call(jsExts(), _ref2) < 0) {
             return next();
           }
           if (depPath === filePath) {
@@ -500,7 +502,7 @@
         });
       };
       return this.readFile(filePath, flags, function(err, fileChanged) {
-        var command, directive, relPath, relPaths, words, _j, _k, _l, _len, _len1, _len2, _ref;
+        var command, directive, relPath, relPaths, words, _j, _k, _l, _len, _len1, _len2, _ref2;
 
         if (err) {
           return callback(err);
@@ -508,9 +510,9 @@
         if (fileChanged) {
           graphChanged = true;
         }
-        _ref = parseDirectives(_this.cache[filePath].data.toString('utf8'));
-        for (_j = 0, _len = _ref.length; _j < _len; _j++) {
-          directive = _ref[_j];
+        _ref2 = parseDirectives(_this.cache[filePath].data.toString('utf8'));
+        for (_j = 0, _len = _ref2.length; _j < _len; _j++) {
+          directive = _ref2[_j];
           words = directive.replace(/['"]/g, '').split(/\s+/);
           command = words[0], relPaths = 2 <= words.length ? __slice.call(words, 1) : [];
           switch (command) {
@@ -607,12 +609,12 @@
       var _this = this;
 
       return this.stat(filePath, flags, function(err, stats) {
-        var data, e, _ref;
+        var data, e, _ref2;
 
         if (err) {
           return callback(err);
         }
-        if (timeEq((_ref = _this.cache[filePath]) != null ? _ref.mtime : void 0, stats.mtime)) {
+        if (timeEq((_ref2 = _this.cache[filePath]) != null ? _ref2.mtime : void 0, stats.mtime)) {
           return callback(null, false);
         }
         if (flags.async) {
@@ -687,54 +689,7 @@
 
   })();
 
-  module.exports.compilers = compilers = {
-    coffee: {
-      match: /\.js$/,
-      compileSync: function(sourcePath, source, useropts) {
-        var compileopts, inbn, inname, inurl, js, opts, outbn, outname, output, outurl, srcmap;
-
-        if (useropts == null) {
-          useropts = {};
-        }
-        opts = {
-          srcmap: false,
-          staticRoot: '',
-          staticRootUrl: '/'
-        };
-        _.extend(opts, useropts);
-        compileopts = {
-          filename: sourcePath
-        };
-        if (opts.srcmap) {
-          outname = "" + sourcePath + ".js";
-          inname = sourcePath;
-          inurl = getUrlPath(inname, opts.staticRoot, opts.staticRootUrl);
-          outurl = getUrlPath(outname, opts.staticRoot, opts.staticRootUrl);
-          inbn = path.basename(inurl);
-          outbn = path.basename(outurl);
-          _.extend(compileopts, {
-            filename: outbn,
-            sourceMap: true,
-            generatedFile: "" + (stripExt(outurl)) + ".min.js",
-            sourceFiles: [inurl]
-          });
-        }
-        output = CoffeeScript.compile(source, compileopts);
-        if (opts.srcmap) {
-          srcmap = output.v3SourceMap;
-          if (_.isString(srcmap)) {
-            srcmap = JSON.parse(srcmap);
-          }
-          js = output.js;
-          return {
-            js: js,
-            srcmap: srcmap
-          };
-        }
-        return output;
-      }
-    }
-  };
+  module.exports.compilers = compilers;
 
   EXPLICIT_PATH = /^\/|:/;
 
@@ -801,118 +756,16 @@
     return _results;
   };
 
-  stripExt = function(filePath) {
-    var _ref;
-
-    if (_ref = path.extname(filePath), __indexOf.call(jsExts(), _ref) >= 0) {
-      return filePath.slice(0, filePath.lastIndexOf('.'));
-    } else {
-      return filePath;
-    }
-  };
-
-  jsExts = function() {
-    var ext;
-
-    return ((function() {
-      var _results;
-
-      _results = [];
-      for (ext in compilers) {
-        _results.push("." + ext);
-      }
-      return _results;
-    })()).concat('.js');
-  };
-
-  minify = function(js, useropts) {
-    var cmpd, inbn, inurl, opts, outbn, outurl, parseopts, sm, smopts, srcmap, stream, streamopts, top;
-
-    if (useropts == null) {
-      useropts = {};
-    }
-    opts = {
-      mangle: false,
-      srcmap: false,
-      outname: '',
-      inname: '',
-      staticRoot: '',
-      staticRootUrl: '/'
-    };
-    _.extend(opts, useropts);
-    parseopts = {};
-    if (opts.inname != null) {
-      parseopts.filename = opts.inname;
-      if ((opts.srcmap != null) && opts.srcmap !== false) {
-        inurl = getUrlPath(opts.inname, opts.staticRoot, opts.staticRootUrl);
-        outurl = getUrlPath(opts.outname, opts.staticRoot, opts.staticRootUrl);
-        inbn = path.basename(inurl);
-        outbn = path.basename(outurl);
-        parseopts.filename = inurl;
-      }
-    }
-    top = uglify.parse(js, parseopts);
-    top.figure_out_scope();
-    cmpd = top.transform(uglify.Compressor({
-      warnings: false
-    }));
-    cmpd.figure_out_scope();
-    if (opts.mangle) {
-      cmpd.mangle_names();
-      cmpd.figure_out_scope();
-    }
-    streamopts = {};
-    if (opts.srcmap !== false) {
-      smopts = {
-        file: "" + (stripExt(outurl)) + ".min.js"
-      };
-      if (opts.srcmap !== true) {
-        smopts.orig = opts.srcmap;
-      }
-      sm = uglify.SourceMap(smopts);
-      streamopts.source_map = sm;
-    }
-    stream = uglify.OutputStream(streamopts);
-    cmpd.print(stream);
-    js = stream.toString();
-    if ((opts.srcmap != null) && opts.srcmap !== false) {
-      srcmap = sm.toString();
-      if (_.isString(srcmap)) {
-        srcmap = JSON.parse(srcmap);
-      }
-      if (opts.srcmap === true) {
-        srcmap.sources = [inurl];
-      }
-      return {
-        js: js,
-        srcmap: srcmap
-      };
-    }
-    return js;
-  };
-
-  getUrlPath = function(absPath, absStaticRoot, staticRootUrl) {
-    absPath = path.resolve(path.normalize(absPath));
-    absStaticRoot = path.resolve(path.normalize(absStaticRoot));
-    if (absStaticRoot[absStaticRoot.length - 1] !== '/') {
-      absStaticRoot += '/';
-    }
-    if (staticRootUrl[staticRootUrl.length - 1] !== '/') {
-      staticRootUrl += '/';
-    }
-    return absPath.replace(absStaticRoot, staticRootUrl);
-  };
-
   sourceMapCat = function(opts) {
-    var combinedGeneratedLine, generated, original, originalLastLine, _i, _len, _original, _ref;
+    var combinedGeneratedLine, generated, original, originalLastLine, _i, _len, _original, _ref2;
 
     generated = new SourceMap.SourceMapGenerator({
       file: opts.filename
     });
     combinedGeneratedLine = 1;
-    _ref = opts.maps;
-    for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-      _original = _ref[_i];
+    _ref2 = opts.maps;
+    for (_i = 0, _len = _ref2.length; _i < _len; _i++) {
+      _original = _ref2[_i];
       if ((_original.empty != null) && _original.empty === true) {
         combinedGeneratedLine += _original.numLines;
         continue;
@@ -943,10 +796,6 @@
       combinedGeneratedLine += originalLastLine;
     }
     return JSON.parse(generated.toString());
-  };
-
-  timeEq = function(date1, date2) {
-    return (date1 != null) && (date2 != null) && date1.getTime() === date2.getTime();
   };
 
 }).call(this);
